+++ tests/test_parsers.py
@@ -1,7 +1,13 @@
-from unittest import TestCase
+from unittest import mock, TestCase
 
 from framewirc.message import build_message, ReceivedMessage
-from framewirc.parsers import is_channel, nick, privmsg
+from framewirc.parsers import (
+    apply_kwargs_parser,
+    apply_message_parser,
+    is_channel,
+    nick,
+    privmsg,
+)
 
 
 class TestIsChannel(TestCase):
@@ -74,11 +80,12 @@ class TestNick(TestCase):
 
 
 class TestPrivmsg(TestCase):
-    sender = b'sender!ident@hostname'
-
-    def processed_message(self, target=b'#target', sender=b'nick!ident@host'):
+    def processed_message(
+            self,
+            target=b'#target',
+            sender=b'nick!ident@host',
+            body='message body'):
         """Build a PRIVMSG, and process it with parsers.privmsg."""
-        body = 'message body'
         message = build_message('PRIVMSG', target, prefix=sender, suffix=body)
         return privmsg(ReceivedMessage(message))
 
@@ -121,3 +128,79 @@ class TestPrivmsg(TestCase):
         result = self.processed_message(target=target, sender=sender)
 
         self.assertEqual(result['channel'], sender_nick)
+
+    def test_not_third_person(self):
+        """Normal messages should not be marked as `third_person`."""
+        result = self.processed_message()
+
+        self.assertFalse(result['third_person'])
+
+    def test_third_person(self):
+        """
+        Set `third_person` when the message is sent as a "CTCP ACTION".
+
+        This is generally the result of a user typing a "/me" command into
+        their client.
+        """
+        result = self.processed_message(body='\1ACTION is third person!\1')
+
+        self.assertEqual(result['raw_body'], b'is third person!')
+        self.assertTrue(result['third_person'])
+
+
+def parser_taking_message(message):
+    return {'key': 'value'}
+
+
+class TestApplyMessageParser(TestCase):
+    def setUp(self):
+        self.client = object()
+        self.handler = mock.Mock()
+        self.message = object()
+
+    def test_result_passed(self):
+        """Dictionary returned from parser passed as kwargs."""
+        wrapped = apply_message_parser(parser_taking_message)(self.handler)
+
+        wrapped(client=self.client, message=self.message)
+
+        self.handler.assert_called_once_with(
+            client=self.client,
+            message=self.message,
+            key='value',
+        )
+
+    def test_parser_called_with_kwarg(self):
+        """
+        The parser is called with `message=message`.
+
+        It's not really a big deal, but it's slightly more flexible if the
+        parser is called with kwargs.
+        """
+        parser = mock.Mock(return_value={})
+        wrapped = apply_message_parser(parser)(self.handler)
+        wrapped(client=self.client, message=self.message)
+        parser.assert_called_once_with(message=self.message)
+
+
+def parser_taking_kwargs(client, message, **kwargs):
+    return {'key': 'value'}
+
+
+class TestApplyKwargsParser(TestCase):
+    def setUp(self):
+        self.client = object()
+        self.handler = mock.Mock()
+        self.message = object()
+
+    def test_result_passed(self):
+        """Dictionary returned from parser passed as kwargs."""
+        wrapped = apply_kwargs_parser(parser_taking_kwargs)(self.handler)
+
+        wrapped(client=self.client, message=self.message)
+
+        self.handler.assert_called_once_with(
+            client=self.client,
+            message=self.message,
+            key='value',
+        )